Extract RoutesAndLinesLists as own component
