+++ src/components/RoutesAndLinesPage.tsx
@@ -1,31 +1,17 @@
-import { useContext } from 'react';
+import React, { useContext } from 'react';
 import { useTranslation } from 'react-i18next';
 import { MapEditorContext } from '../context/MapEditor';
 import { ModalMapContext } from '../context/ModalMapContext';
-import {
-  RouteLine,
-  RouteRoute,
-  useListChangingRoutesQuery,
-  useListOwnLinesQuery,
-} from '../generated/graphql';
-import { mapListChangingRoutesResult, mapListOwnLinesResult } from '../graphql';
 import { Container, Row } from '../layoutComponents';
 import { Path, routes } from '../routes'; // eslint-disable-line import/no-cycle
 import { SimpleButton } from '../uiComponents';
-import { LineTableRow } from './LineTableRow'; // eslint-disable-line import/no-cycle
-import { RoutesTable } from './RoutesTable';
-import { RoutesTableRow } from './RoutesTableRow'; // eslint-disable-line import/no-cycle
+import { RoutesAndLinesLists } from './routes-and-lines/RoutesAndLinesLists'; // eslint-disable-line import/no-cycle
 
 export const RoutesAndLinesPage = (): JSX.Element => {
   const { t } = useTranslation();
   const { dispatch: modalMapDispatch } = useContext(ModalMapContext);
   const { dispatch: mapEditorDispatch } = useContext(MapEditorContext);
   const createLineReactRoute = routes[Path.createLine];
-  const changingRoutesResult = useListChangingRoutesQuery();
-  const changingRoutes = mapListChangingRoutesResult(changingRoutesResult);
-  const ownLinesResult = useListOwnLinesQuery();
-  const ownLines = mapListOwnLinesResult(ownLinesResult);
-
   const onOpenModalMap = () => {
     mapEditorDispatch({ type: 'reset' });
     modalMapDispatch({ type: 'open' });
@@ -46,24 +32,7 @@ export const RoutesAndLinesPage = (): JSX.Element => {
           {t('lines.createNew')}
         </SimpleButton>
       </Row>
-      <h2 className="mb-14 mt-12 text-4xl font-bold">
-        {t('routes.changingRoutes')}
-      </h2>
-      {changingRoutes && changingRoutes.length > 0 && (
-        <RoutesTable>
-          {changingRoutes.map((item: RouteRoute) => (
-            <RoutesTableRow key={item.route_id} route={item} />
-          ))}
-        </RoutesTable>
-      )}
-      <h2 className="mb-14 mt-12 text-4xl font-bold">{t('routes.ownLines')}</h2>
-      {ownLines && ownLines.length > 0 && (
-        <RoutesTable>
-          {ownLines.map((item: RouteLine) => (
-            <LineTableRow key={item.line_id} line={item} />
-          ))}
-        </RoutesTable>
-      )}
+      <RoutesAndLinesLists />
     </Container>
   );
 };
