+++ src/components/routes-and-lines/RoutesList.tsx
@@ -1,0 +1,16 @@
+import React from 'react';
+import { RouteRoute } from '../../generated/graphql';
+import { RoutesTable } from '../RoutesTable';
+import { RoutesTableRow } from '../RoutesTableRow'; // eslint-disable-line import/no-cycle
+
+type Props = {
+  routes?: RouteRoute[];
+};
+
+export const RoutesList = ({ routes }: Props): JSX.Element => (
+  <RoutesTable>
+    {routes?.map((item: RouteRoute) => (
+      <RoutesTableRow key={item.route_id} route={item} />
+    ))}
+  </RoutesTable>
+);
